:+1: Expose only types to improve backward compatibility
